added --weight option for refmac refinement
after this comment from Paul McEwan:
There is only one comment I would make and this is in references to the refinement. Although the automatic geometry weight appears to be a bit low as judges by some distortions of a few residues in weaker density, but this may a product of the low resolution of this particular data.